Refactor: Implement interface for ReviewerController (reviewer)
Extracted an interface `IReviewerController` from the
`ReviewerControllerImpl` class.  Renamed
`ReviewerControllerImpl` to `ReviewerController` and
implemented the new interface.

Also renamed `UserService` to `IUserService` and updated
`UserServiceImpl` to implement the new interface.

This change improves code structure and allows for easier
testing and extension in the future.  Deleted the old
`ReviewerControllerImpl` class.
