+++ src/main/java/konkuk/ptal/service/IUserService.java
@@ -4,7 +4,7 @@ package konkuk.ptal.service;
 import konkuk.ptal.dto.request.CreateReviewerRequestDto;
 import konkuk.ptal.entity.Reviewer;
 
-public interface UserService {
+public interface IUserService {
     Reviewer registerReviewer(CreateReviewerRequestDto dto, Long authenticatedUserId);
 
     Reviewer getReviewer(Long id);
